Refactor order blueprint import in routes/__init__.py
