+++ routes/__init__.py
@@ -1,6 +1,6 @@
 from .user import user_bp
 from .product import product_bp
-from .order import order_bp
+from .push import order_bp
 
 # Blueprintをリストとしてまとめる
 blueprints = [
